allow float for days

+++ src/task/task.entity.ts
@@ -19,7 +19,7 @@ export class Task {
   @Column({ type: 'text', name: 'description' })
   description!: string;
 
-  @Column({ type: 'int', name: 'estimated_days' })
+  @Column({ type: 'float', name: 'estimated_days' })
   estimatedDays!: number;
 
   @Column({ type: 'boolean', name: 'is_completed', default: false })
